// types.ts
export interface TagData {
  id: number;
  productDescription: string;
  brand: string;
  code: string;
  installmentsCount: string;
  installmentsText: string;
  installmentsPrice: string;
  listPriceLabel: string;
  listPrice: string;
  validityDate: string;
}

export type LayoutMode = '1x' | '2x' | '4x';